pyproject.toml
[tool.poetry]
name = "doc-modernizer"
version = "0.1.0"
description = "A tool to modernize legacy Word work instructions"
authors = ["Evan Stachowiak <estachowiak@styberg.com>"]
license = "MIT"

[[tool.poetry.packages]]
include = "modernizer"

[tool.poetry.dependencies]
python = "^3.10"
python-docx = "^1.1.2"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
pre-commit = "^3.0"
black = "^24.0"
# ruff removed temporarily

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
